Allow user to specify location of Decklink SDK to ./configure
Example usage:

./configure --with-decklink-sdk=/home/devin/decklink-sdk

This is needed in cases where the bundled SDK headers won't work,
such as when building on OSX.

If the option isn't specified, use the bundled headers, but only
if on Linux.  The bundled headers won't work on any other platform,
so explicitly error out rather than using headers which won't work.