Modify string routines to work on platforms other than Linux
The GetString() method has different variable types depending on
the platform, so introduce some portability macros that use the
native variable type for the given platform, and then convert the
result into a (char *) so that the application works as expected.

Note: only tested with OSX.  It's likely additional changes to
the macros would be needed for other platforms such as Windows.