Introduce ability to deserialize an SCTE-104 payload without VANC
The existing parse routine only supported taking in an SMPTE 2010
packet within a VANC ADF sequence.  Add a new parsing function which
takes in just the SCTE-104 payload, and refactor the parsing into
common code that can be shared by both the VANC SMPTE 2010 input
and SCTE-104 payload input use cases.

This patch also addresses a potential illegal memory access issue
where the parser was blindly copying 200 bytes of data into the
payload field.  This could result in an access to unallocated
memory if the packet is actually smaller than 200 bytes and near
the end of the VANC line.  We now use the actual size of the VANC
packet as the basis for copying into the payload field.