Fix cut/paste error in minutes field
The deserialization of the tens-of-minutes field was referencing the
values from the tens-of-seconds field as a result of a cut/paste
error.  It passed the test suite out of dumb luck because the one
test case we had happened to have those bits set for both the
minutes and seconds field.