Add public function to get VANC line numbers for various video modes
There are use cases where we need to know which lines contain VANC,
in particular for interlaced video modes.  We also want to know
where the switching line is for each mode to assist protocols where
the preferred line is defined in the spec as "switchingline + 2".

Create a public function which provides this info for all known
video modes.